# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AnimeTower API Documentation",
        "description": "API Documentation for AnimeTower Application",
        "contact": {
            "name": "Support Team",
            "email": "mahmoud@7modo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/v1/anime": {
            "get": {
                "tags": [
                    "Anime"
                ],
                "summary": "List all anime",
                "operationId": "90bea2738456727c6049c238aa404252",
                "responses": {
                    "200": {
                        "description": "List of anime retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Anime"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Anime"
                ],
                "summary": "Create a new anime",
                "operationId": "7654a11c35f261cc06583a67b53d48fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "synopsis",
                                    "type",
                                    "status",
                                    "genre_ids"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Naruto"
                                    },
                                    "synopsis": {
                                        "type": "string",
                                        "example": "A young ninja seeks to become the leader of his village"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "TV",
                                            "Movie",
                                            "OVA"
                                        ],
                                        "example": "TV"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "ongoing",
                                            "completed"
                                        ],
                                        "example": "ongoing"
                                    },
                                    "episodes": {
                                        "type": "integer",
                                        "example": 220
                                    },
                                    "genre_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1
                                        ]
                                    },
                                    "cover_image": {
                                        "type": "string",
                                        "example": "naruto.jpg"
                                    },
                                    "trailer_url": {
                                        "type": "string",
                                        "example": "@https://example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Anime created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Anime created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Anime"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/anime/{id}": {
            "get": {
                "tags": [
                    "Anime"
                ],
                "summary": "Get anime details",
                "operationId": "0feac4ec5d854f8a47947e37a7a08ac1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Anime details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Anime"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Anime not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Anime"
                ],
                "summary": "Update anime details",
                "operationId": "8f96a0ee73ccbf0a43b195de952f3697",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "synopsis": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "episodes": {
                                        "type": "integer"
                                    },
                                    "genre_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "cover_image": {
                                        "type": "string"
                                    },
                                    "trailer_url": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Anime updated successfully"
                    },
                    "404": {
                        "description": "Anime not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Anime"
                ],
                "summary": "Delete an anime",
                "operationId": "bac6130d22fbf9e0e7bf0ed5a942a265",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Anime deleted successfully"
                    },
                    "404": {
                        "description": "Anime not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Anime": {
                "required": [
                    "title",
                    "slug",
                    "synopsis"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Naruto"
                    },
                    "slug": {
                        "type": "string",
                        "example": "naruto"
                    },
                    "synopsis": {
                        "type": "string",
                        "example": "A young ninja seeks to become the leader of his village"
                    },
                    "type": {
                        "type": "string",
                        "example": "TV"
                    },
                    "status": {
                        "type": "string",
                        "example": "ongoing"
                    },
                    "episodes": {
                        "type": "integer",
                        "example": 220
                    },
                    "cover_image": {
                        "type": "string",
                        "example": "naruto.jpg"
                    },
                    "trailer_url": {
                        "type": "string",
                        "example": "https://youtube.com/watch?v=..."
                    },
                    "score": {
                        "type": "number",
                        "format": "float",
                        "example": 4.5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}